docs(class and requirements): Code skeleton
Added the code skeleton for the trader class.
I target a minimal code structure and autnomous
class operation. Will be able to declare multiple trader
that understand the market and operate based on specific
rules (agressive, short-seller, optimistic, and so on...)

Will be able to give trading strategy with natural language and it
will understand how it should operate, and get feedback with
Reinforcment Learning